requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-bulk-update==2.2.0
django-filter==23.2
djangorestframework==3.14.0
idna==3.4
Markdown==3.4.3
phonenumbers==8.13.14
Pillow==9.5.0
pytz==2023.3
reportlab==4.0.4
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==2.0.3
